rename styles.tsx to widgetWrapper.tsx
