+++ src/sentry/static/sentry/app/views/dashboardsV2/widgetWrapper.tsx
@@ -3,7 +3,7 @@ import {motion} from 'framer-motion';
 
 import {Widget} from './types';
 
-export const WidgetWrapper = styled(motion.div)<{displayType: Widget['displayType']}>`
+const WidgetWrapper = styled(motion.div)<{displayType: Widget['displayType']}>`
   position: relative;
   /* Min-width prevents grid items from stretching the grid */
   min-width: 200px;
@@ -18,3 +18,5 @@ export const WidgetWrapper = styled(motion.div)<{displayType: Widget['displayTyp
     }
   }};
 `;
+
+export default WidgetWrapper;
